+++ tests/functional/test_cli.py
@@ -457,8 +457,7 @@ class TestWeaverClient(TestWeaverClientBase):
 
         result = mocked_sub_requests(self.app, self.client.undeploy, other_process)
         assert result.success
-        assert result.body.get("undeploymentDone", None) is True
-        assert "undefined" not in result.message
+        assert not result.body
 
         path = f"/processes/{other_process}"
         resp = mocked_sub_requests(self.app, "get", path, expect_errors=True)